setting up all flags with yaml file
